undoDelete: added Snackbar to enable undo delete action

+++ app/src/main/java/com/example/wordsmemory/presentation/fragment/vocabulary/words/VocabularyWordsFragment.kt
@@ -16,10 +16,13 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.wordsmemory.R
 import com.example.wordsmemory.presentation.adpater.VocabularyItemAdapter
 import com.example.wordsmemory.databinding.VocabularyWordsFragmentBinding
+import com.example.wordsmemory.presentation.adpater.CategoryItemAdapter
 import com.example.wordsmemory.presentation.helper.SwipeToDeleteCallback
 import com.example.wordsmemory.presentation.fragment.vocabulary.VocabularyFragment
 import com.example.wordsmemory.presentation.fragment.vocabulary.VocabularyFragmentDirections
 import com.example.wordsmemory.presentation.fragment.vocabulary.category.CategoryFragmentDirections
+import com.google.android.material.snackbar.BaseTransientBottomBar
+import com.google.android.material.snackbar.Snackbar
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.InternalCoroutinesApi
@@ -111,7 +114,20 @@ class VocabularyWordsFragment(private val _categoryId: Int = -1) : Fragment() {
     private fun setSwipeGesture() {
         val swipeHandler = object : SwipeToDeleteCallback(requireContext()) {
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                _viewModel.deleteItem((viewHolder as VocabularyItemAdapter.VocabularyItemViewHolder).itemId)
+                var delete = true
+                Snackbar.make(_binding.coordinatorLayout, R.string.deleted, Snackbar.LENGTH_LONG)
+                    .setAction(R.string.undo) { delete = false }
+                    .addCallback(object : BaseTransientBottomBar.BaseCallback<Snackbar>() {
+                        override fun onDismissed(transientBottomBar: Snackbar?, event: Int) {
+                            super.onDismissed(transientBottomBar, event)
+                            if (delete) {
+                                _viewModel.deleteItem((viewHolder as VocabularyItemAdapter.VocabularyItemViewHolder).itemId)
+                            } else {
+                                _binding.vocabularyList.adapter?.notifyItemChanged(viewHolder.layoutPosition)
+                            }
+                        }
+                    })
+                    .show()
             }
         }
         val itemTouchHelper = ItemTouchHelper(swipeHandler)
